# app/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Blogg Application API",
      "description": "Node.js Blogg Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000/",
    "basePath": "/",
...
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "test" :{
        "get" :{
          "tags":[
            "Tests"
          ],
          "parameters": [
            {
              "name": "x-access-token",
              "in": "header",
              "description": "Json Web Token",              
              "type" : "string",
              "required" : "true"
            }
          ],
          "summary" : "Get all Tests",
          "responses" :{
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Tests"
              }
            }
          }
        },
        "post": {
          "tags": [
            "Tests"
          ],
          "summary" : "Create New Tests",
          "description": "Create New Tests",
          "parameters": [
              {
                "name": "x-access-token",
                "in": "header",
                "type" : "string",
                "required" : "true"
              },
            {
              "name": "test",
              "in": "body",
              "description": "Test that we want to create",
              "schema": {
                "$ref": "#/definitions/Test"
              }
            }
          ],
          "produces": [
            "application/json"
          ]
        }
      },
...
    "employee" : {
      "get" :{
        "tags":[
          "Employee"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Json Web Token",              
            "type" : "string",
            "required" : "true"
          }
        ],
        "summary" : "Get all Employees",
        "responses" :{
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Employees"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employee"
        ],
        "summary" : "Create New Employee",
        "description": "Create New Employee",
        "parameters": [
            {
              "name": "x-access-token",
              "in": "header",
              "type" : "string",
              "required" : "true"
            },
          {
            "name": "employee",
            "in": "body",
            "description": "Employee that we want to create",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "produces": [
          "application/json"
        ]
      }
    },
...
    },
    "definitions": {
...
      "Test" :{
        "required": [
          "name",
          "number"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "number": {
            "type": "number"
          }
        }
      },
      "Tests" :{
        "type": "array",
        "$ref": "#/definitions/Test"
      },
...
       "Employee" : {
        "properties" : {
          "id": {
           "type": "number"
         },
         "firstname": {
           "type": "string"
         },
         
         "lastname" : {
          "type": "string"
         },
         "email" : {
          "type": "string"
         },
         "contact" : { "type" : "number"},
         "city_id" : {"type" : "number"},
         "state_id" : {"type" : "number"},
         "country_id" : {"type" : "number"},
         "dateofbirth" : {"type" : "string"},
         "dateofjoining" : {"type" : "string"},
         "salary" : {"type" : "number"},
         "gender" : { "type": "string" },
         "depertment_id" : {"type" : "number"},
         "isactive" : {"type" : "boolean"},
          "skills" : {"type" : "string"},
          "summary" : {"type" : "string"},
          "experience" : {"type" :"number"}
        }
       },
       "Employees" : {
        "type": "array",
        "$ref": "#/definitions/Employee"
       }
    }
  }